+++ src/main/java/uz/pdp/dto/SellerRequestDto.java
@@ -1,6 +1,6 @@
 package uz.pdp.dto;
 
-import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,6 +9,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class SellerRequestDto {
-    @NotBlank(message = "User ID is required")
+    @NotNull(message = "User ID cannot be null")
     private Long userId;
 }